ASC-8 Added another module for telescope and enviroment parameters (eta_pol is the only param remaining in instrument setup params)
